major internal rewrite
